ãV3 마마ë£무리

+++ src/main/java/com/tenco/blog/board/BoardController.java
@@ -1,5 +1,7 @@
 package com.tenco.blog.board;
 
+import com.tenco.blog.user.User;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -18,7 +20,6 @@ import java.util.List;
 public class BoardController {
 
     // DI처리
-    private final BoardNativeRepository boardNativeRepository;
     private final BoardPersistRepository boardPersistRepository;
 
     /**
@@ -29,7 +30,14 @@ public class BoardController {
      */
     // 자원의 요청 , 화면의 요청
     @GetMapping("/board/save-form")
-    public String saveForm() {
+    public String saveForm(HttpSession httpSession) {
+        // 로그인 여부 체크 - 즉 로그인 한 사용자만 이 페이지 안에 들어 수 있음
+        // 1. 인증 검사
+        User sessionUser = (User) httpSession.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
+
         return "board/save-form";
     }
 
@@ -44,14 +52,36 @@ public class BoardController {
 
     //자원의 생성 ,기능의 요청 - 백그라운드에서 동작하는가?
     @PostMapping("/board/save")// - 알아서 구분 가능함
-    public String saveProc(BoardRequest.SaveDTO saveDTO) {
-        // 이제 따로 작성해서 대입할 필요없이 saveDTO를 통해 만들어둔 객체를
-        // 이용해서 저장이 가능하다.
-        boardPersistRepository.save(saveDTO.toEntity());
+    public String saveProc(BoardRequest.SaveDTO saveDTO, HttpSession session) {
 
-        // redirect : 다시 URL 요청 해!
-        // return "redirect:/";
-        return "redirect:/";
+        log.info("===게시글 저장 요청====");
+        // 이 요청 사용자가 로그인을했다면 로그인 정보를 세션 메모리에서 가져오면된다.
+        // 1. 세션에서 로그인한 사용자 정보 가져오기
+        // HttpSession으로 저장한 변수 호출 - 가져오기
+        User sessionUser = (User) session.getAttribute("sessionUser");
+
+        // 2. 로그인 여부 확인
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
+
+        try {
+            //3. 로그인 된 사용자
+            //3.1 유효성 검사
+            saveDTO.validate();
+
+            Board board = saveDTO.toEntity(sessionUser);
+            boardPersistRepository.save(board);
+
+            // redirect : 다시 URL 요청 해!
+            // return "redirect:/";
+            return "redirect:/";
+        } catch (Exception e) {
+            System.out.println("에러 발생 : " + e.getMessage());
+            return "board/save-form";
+        }
+
+
     } // end of saveProc
 
     /**
@@ -88,23 +118,54 @@ public class BoardController {
         return "board/detail";
     } // end of detailPage
 
-    // 게시글 제거
+    // 게시글 삭제
+    // 1. 로그인 여부 확인
+    // 2. 삭제할 게시글이 본인이 작성한 게시글인지 확인 (권한 확인,인가 처리)
+    // 3. 인가처리 후  삭제 진행
     @PostMapping("/board/{id}/delete")
-    public String deleteProc(@PathVariable(name = "id") Integer id) {
+    public String deleteProc(@PathVariable(name = "id") Integer id, HttpSession session) {
         //boardNativeRepository.deleteById(id);
-        boardPersistRepository.deleteById(id);
+
+        log.info("=== 게시글 삭제 요청 ===");
+        //인증검사
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
+        try {
+            // 삭제할 게시글 조회 (권한 체크 = 인가 처리)
+            Board board = boardPersistRepository.findById(id);
+            if (board.getUser().getId() == sessionUser.getId()) {
+                boardPersistRepository.deleteById(id);
+            }
+
+
+        } catch (Exception e) {
+            return "redirect:/";
+        }
+
         // PRG 패턴 (Post -> Redirect -> Get) 적용
         return "redirect:/";
     } // end of deleteProc
 
-    //게시글 수정 시 --> 다시 사용자가 게시글 작성할 수 있도록 설계
+    //게시글 수정 시 --> 사용자에게 해당 게시물 내용을 보여줘야 한다.
     // http://localhost:8080/board/1/update-form
     @GetMapping("/board/{id}/update-form")
-    public String updateFormPage(@PathVariable(name = "id") Integer id, Model model) {
-        // 사용자에게 해당 게시물 내용을 보여줘야 한다.
-        // 조회 기능 - 게시글 id로
+    public String updateFormPage(@PathVariable(name = "id") Integer id, Model model, HttpSession session) {
 
+        // 인증 처리
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
+
+        // 인가 처리
+        // 조회 기능 - 게시글 id로
         Board board = boardPersistRepository.findById(id);
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new RuntimeException("수정 권한이 없습니다.");
+        }
+
         model.addAttribute("board", board);
 
         return "board/update-form";
@@ -113,15 +174,30 @@ public class BoardController {
     // 게시글 수정
     @PostMapping("/board/{id}/update")
     // 메세지 컨버터 객체가 동작해서 자도으로 객체를 생성하고 값을 매핑해준다.(뷰 리졸브느낌)
-    public String updateProc(@PathVariable(name = "id") Integer id, BoardRequest.UpdateDTO updateDTO) {
+    public String updateProc(@PathVariable(name = "id") Integer id, BoardRequest.UpdateDTO updateDTO
+            , HttpSession session) {
         //@PathVariable - 경로 변수를 가져올 수 있다.
 
-        // 유효성 검사 username,title,content 유효성 검사
-        updateDTO.validate();
+        // 인증 검사
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            return "redirect:/login-form";
+        }
 
-        // 2, DAO 계층으로 전달
-        boardPersistRepository.updateById(id, updateDTO);
+        try {
+            // 유효성 검사
+            updateDTO.validate();
+            // 인가 처리
+            Board board = boardPersistRepository.findById(id);
+            if (sessionUser.getId() != board.getUser().getId()) {
+                throw new RuntimeException("수정할 권한이 없습니다.");
+            }
+            // 2, DAO 계층으로 전달
+            boardPersistRepository.updateById(id, updateDTO);
 
+        } catch (Exception e) {
+            return "redirect:/board/" + id + "/update-form";
+        }
 
         return "redirect:/board/" + id;
     } // end of updateProc
